requirements.txt
eth-brownie==1.18.1
eciespy==0.3.11
httpx==0.22.0
numpy==1.21.0
PyNaCl==1.5.0
python-dotenv==0.16.0
scikit-learn==1.0.1
web3==5.27.0

# Ocean integration
ocean-lib==0.5.0
# Fileprovider
fastapi==0.75.0
uvicorn
python-multipart